requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
joblib>=1.0.1
flask>=2.0.1
python-dateutil>=2.8.2
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
gevent>=22.10.2 